PI-10228 Pages: Product-List configuration
